requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
markdown==3.7
python-frontmatter==1.1.0
itsdangerous==2.2.0
pygments==2.18.0
